+++ CacheDatabaseQueriesApiSample/Startup.cs
@@ -1,5 +1,4 @@
-﻿using LazyCache;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -30,7 +29,6 @@ namespace CacheDatabaseQueriesApiSample
 
             // Register IAppCache as a singleton CachingService
             services.AddLazyCache();
-
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.