+++ LazyCache/Mocks/MockCacheEntry.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.Primitives;
+
+namespace LazyCache.Mocks
+{
+    public class MockCacheEntry : ICacheEntry
+    {
+        public MockCacheEntry(string key)
+        {
+            Key = key;
+        }
+
+        public void Dispose()
+        {
+        }
+
+        public object Key { get; }
+        public object Value { get; set; }
+        public DateTimeOffset? AbsoluteExpiration { get; set; }
+        public TimeSpan? AbsoluteExpirationRelativeToNow { get; set; }
+        public TimeSpan? SlidingExpiration { get; set; }
+        public IList<IChangeToken> ExpirationTokens { get; }
+        public IList<PostEvictionCallbackRegistration> PostEvictionCallbacks { get; }
+        public CacheItemPriority Priority { get; set; }
+        public long? Size { get; set; }
+    }
+}