+++ LazyCache/Mocks/MockCachingService.cs
@@ -11,52 +11,25 @@ namespace LazyCache.Mocks
     public class MockCachingService : IAppCache
     {
         public ICacheProvider CacheProvider { get; } = new MockCacheProvider();
-
-        public void Add<T>(string key, T item)
-        {
-        }
-
-        public void Add<T>(string key, T item, DateTimeOffset expires)
-        {
-        }
+        public CacheDefaults DefaultCachePolicy { get; set; }
 
         public T Get<T>(string key)
         {
             return default(T);
         }
 
-        public T GetOrAdd<T>(string key, Func<T> addItemFactory)
+        public T GetOrAdd<T>(string key, Func<ICacheEntry, T> addItemFactory)
         {
-            return addItemFactory.Invoke();
-        }
-
-        public T GetOrAdd<T>(string key, Func<T> addItemFactory, DateTimeOffset expires)
-        {
-            return addItemFactory.Invoke();
+            return addItemFactory(new MockCacheEntry(key));
         }
 
         public void Remove(string key)
         {
         }
 
-        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, MemoryCacheEntryOptions policy)
+        public Task<T> GetOrAddAsync<T>(string key, Func<ICacheEntry, Task<T>> addItemFactory)
         {
-            return addItemFactory.Invoke();
-        }
-
-        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory)
-        {
-            return addItemFactory.Invoke();
-        }
-
-        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, DateTimeOffset expires)
-        {
-            return addItemFactory.Invoke();
-        }
-
-        public Task<T> GetOrAddAsync<T>(string key, Func<Task<T>> addItemFactory, TimeSpan slidingExpiration)
-        {
-            return addItemFactory.Invoke();
+            return addItemFactory(new MockCacheEntry(key));
         }
 
         public Task<T> GetAsync<T>(string key)
@@ -64,52 +37,7 @@ namespace LazyCache.Mocks
             return Task.FromResult(default(T));
         }
 
-
-        public void Add<T>(string key, T item, TimeSpan slidingExpiration)
-        {
-        }
-
         public void Add<T>(string key, T item, MemoryCacheEntryOptions policy)
-        {
-        }
-
-        public T GetOrAdd<T>(string key, Func<T> addItemFactory, TimeSpan slidingExpiration)
-        {
-            return addItemFactory.Invoke();
-        }
-
-        public T GetOrAdd<T>(string key, Func<T> addItemFactory, MemoryCacheEntryOptions policy)
-        {
-            return addItemFactory.Invoke();
-        }
-    }
-
-    public class MockCacheProvider : ICacheProvider
-    {
-        public void Set(string key, object item, MemoryCacheEntryOptions policy)
-        {
-        }
-
-        public object Get(string key)
-        {
-            return null;
-        }
-
-        public object GetOrCreate<T>(string key, Func<ICacheEntry, T> func)
-        {
-            return func(null);
-        }
-
-        public void Remove(string key)
-        {
-        }
-
-        public Task<T> GetOrCreateAsync<T>(string key, Func<ICacheEntry, Task<T>> func)
-        {
-            return func(null);
-        }
-
-        public void Dispose()
         {
         }
     }