+++ ReleaseNotes.md
@@ -11,9 +11,17 @@
   `Func<ICacheProvider> DefaultCacheProvider { get; }`
   
   By default we use a shared in-memory cache but each instance can have it's underlying cache provider overridden from it's constructor.
-- Make methods on CachingService virtual
+- Make methods on CachingService virtual/protected to enable 
 - Add LazyCache.AspNetCore for dependency injection registration - ServiceCollection.AddLazyCache();
 - Update sample to use aspnet core and LazyCache.AspNetCore
+- New IAppCache.DefaultCachePolicy to replace CachingService.DefaultCacheDuration
+- Moved most CachingService method overloads to extension methods on IAppCache in AppCacheExtensions. API should be backwards compatible but as now extension methods this is technically an API breaking changing.
+- Added new methods on IAppCache to allow you to specify cache expiry options on executution of the item factory
+   
+  `GetOrAdd<T>(string key, Func<ICacheEntry, T> addItemFactory)`
+  
+  `Task<T> GetOrAddAsync<T>(string key, Func<ICacheEntry, Task<T>> addItemFactory)`
+
 
 ## Version 0.7.1
 - Fix async/sync interopability bug, see https://github.com/alastairtree/LazyCache/issues/12